iris_predictor.py
import streamlit as st
import cv2
import numpy as np
from pathlib import Path
import tempfile
from typing import Dict, List, Any, Tuple, Optional


class IrisPredictor:
    """Class for analyzing iris images and generating predictions/queries."""
    
    def __init__(self):
        """Initialize the iris predictor."""
        # In a real implementation, you would load models here
        pass
    
    def preprocess_image(self, image_path: str) -> np.ndarray:
        """
        Preprocess the iris image for analysis.
        
        Args:
            image_path: Path to the iris image
            
        Returns:
            Preprocessed image as numpy array
        """
        # Load image
        img = cv2.imread(image_path)
        if img is None:
            raise ValueError(f"Could not load image from {image_path}")
            
        # Convert to RGB (OpenCV loads as BGR)
        img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
        
        # In a real implementation, you would resize, normalize, etc.
        
        return img
    
    def detect_iris(self, image: np.ndarray) -> Tuple[np.ndarray, Dict[str, Any]]:
        """
        Detect and extract the iris region from an eye image.
        
        Args:
            image: Input eye image
            
        Returns:
            Extracted iris image and metadata
        """
        # Placeholder implementation
        # In a real implementation, you would:
        # 1. Detect the eye
        # 2. Find the iris boundary using Hough transform
        # 3. Extract the iris region
        
        # For now, just return the original image and dummy metadata
        height, width = image.shape[:2]
        
        # Dummy iris circle (center_x, center_y, radius)
        iris_circle = (width // 2, height // 2, min(width, height) // 3)
        
        metadata = {
            "iris_detected": True,
            "iris_circle": iris_circle,
            "confidence": 0.95
        }
        
        return image, metadata
    
    def analyze_iris(self, iris_image: np.ndarray) -> Dict[str, Any]:
        """
        Analyze the iris image and identify health indicators.
        
        Args:
            iris_image: Preprocessed iris image
            
        Returns:
            Dictionary containing analysis results
        """
        # Placeholder implementation
        # In a real implementation, you would:
        # 1. Segment the iris into zones
        # 2. Detect features in each zone
        # 3. Map features to health conditions
        
        # Dummy analysis results
        return {
            "zones": {
                "liver": {"condition": "normal", "confidence": 0.85},
                "kidney": {"condition": "stressed", "confidence": 0.75},
                "lungs": {"condition": "normal", "confidence": 0.90},
                "heart": {"condition": "normal", "confidence": 0.82},
                "stomach": {"condition": "inflamed", "confidence": 0.78},
            },
            "overall_health": "good",
        }
    
    def generate_queries(self, analysis: Dict[str, Any]) -> List[str]:
        """
        Generate natural language queries based on iris analysis.
        
        Args:
            analysis: Results of iris analysis
            
        Returns:
            List of queries related to the analysis
        """
        queries = []
        
        # Generate queries for concerning areas
        for zone, info in analysis["zones"].items():
            if info["condition"] != "normal" and info["confidence"] > 0.7:
                queries.append(f"Why does the iris show {info['condition']} condition in {zone} zone?")
                queries.append(f"How to improve {zone} health when iris shows {info['condition']} signs?")
        
        # Add general query if no specific issues
        if not queries:
            queries.append("What does a healthy iris look like?")
            
        return queries
    
    def process_iris_image(self, image_path: str) -> Dict[str, Any]:
        """
        Process an iris image and return analysis and queries.
        
        Args:
            image_path: Path to the iris image
            
        Returns:
            Dictionary with analysis results and suggested queries
        """
        # Preprocess the image
        preprocessed_img = self.preprocess_image(image_path)
        
        # Detect iris
        iris_img, iris_metadata = self.detect_iris(preprocessed_img)
        
        if not iris_metadata["iris_detected"]:
            return {"error": "No iris detected in the image"}
        
        # Analyze iris
        analysis = self.analyze_iris(iris_img)
        
        # Generate queries
        queries = self.generate_queries(analysis)
        
        return {
            "analysis": analysis,
            "metadata": iris_metadata,
            "queries": queries,
            "image": iris_img
        }
