requirements.txt
streamlit>=1.22.0
pymupdf>=1.21.1
sentence-transformers>=2.2.2
qdrant-client>=1.1.1
numpy>=1.24.3
opencv-python>=4.7.0
pillow>=9.5.0
matplotlib>=3.7.1
pytesseract>=0.3.10
